Working on getting the symlink script working properly
I had an issue where the script must have made a symlink of .local instead of just the files I wanted linked. This was causing all kinds of issues until I found it. Also worked with chatgpt to modify the script so it didn't remove anything unless it has a corresponding file in the dotfiles directory to create a symlink to.